scraper/requirements.txt
mwclient>=0.10.1
beautifulsoup4>=4.9.0
requests>=2.25.0
mwparserfromhell>=0.6.4
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
